// src/models/interface.ts
export type Id = mongoose.Types.ObjectId;
// ...
export type Group =
  | "A"
  | "B"
  | "C"
  | "Dog"
  | "E"
  | "F"
  | "G"
  | "H"
  | "J"
  | "K"
  | "L"
  | "M"
  | "N"
  | "P"
  | "Q"
  | "R"
  | "S"
  | "T";
export type Size = "S" | "M" | "L" | "XL" | "XXL" | "3XL";
export type RoleCamp = Mode | "peto";
export type Role = RoleCamp | "admin";
export type Mode = "nong" | "pee";
// ...
export const choices = ["A", "B", "C", "D", "E"] as const;
export type Choice = (typeof choices)[number];
export interface InterChoiceQuestion {
  campId: Id;
  question: string;
  a: string;
  b: string;
  c: string;
  d: string;
  e: string;
  _id: Id;
  scoreA: number;
  scoreB: number;
  scoreC: number;
  scoreD: number;
  scoreE: number;
  nongAnswerA: number;
  nongAnswerB: number;
  nongAnswerC: number;
  nongAnswerD: number;
  nongAnswerE: number;
  peeAnswerA: number;
  peeAnswerB: number;
  peeAnswerC: number;
  peeAnswerD: number;
  peeAnswerE: number;
  correct: Choice | "-";
  order: number;
  answerIds: Id[];
  //public
}
export interface EditChoiceQuestion {
  _id: Id | null;
  question: string;
  a: string;
  b: string;
  c: string;
  d: string;
  e: string;
  scoreA: number;
  scoreB: number;
  scoreC: number;
  scoreD: number;
  scoreE: number;
  correct: Choice | "-";
  order: number;
  //public
}
export interface InterTextQuestion {
  question: string;
  _id: Id;
  campId: Id;
  answerIds: Id[];
  score: number;
  order: number;
  //public
}
export interface EditTextQuestion {
  question: string;
  _id: Id | null;
  score: number;
  order: number;
  //public
}
export interface InterAnswerContainer {
  choiceAnswerIds: Id[];
  textAnswerIds: Id[];
  campId: Id;
  userId: Id;
  _id: Id;
  role: RoleCamp;
  //private
}
export interface InterTextAnswer {
  _id: Id;
  answer: string;
  userId: Id;
  questionId: Id;
  score: number;
  containerId: Id;
  //private
}
export interface ChoiceAnswerPack {
  answer: Choice | "-";
  questionId: Id;
  answerId: Id | null;
  //private
}
export interface TextAnswerPack {
  answer: string;
  questionId: Id;
  answerId: Id | null;
  //private
}
export interface AnswerPack {
  campId: Id;
  textAnswers: TextAnswerPack[];
  choiceAnswers: ChoiceAnswerPack[];
  //private
}
export interface EditQuestionPack {
  campId: Id;
  texts: EditTextQuestion[];
  choices: EditChoiceQuestion[];
  //public
}
export interface GetChoiceQuestion extends InterChoiceQuestion {
  answer: Choice | "-";
  answerId: Id | null;
  //private
}
export interface GetTextQuestion extends InterTextQuestion {
  answer: string;
  answerId: Id | null;
  answerScore: number;
  //private
}
export interface GetAllQuestion {
  choices: GetChoiceQuestion[];
  texts: GetTextQuestion[];
  canAnswerTheQuestion: boolean;
  //private
}
export interface UserAndAllQuestionPack {
  user: BasicUser;
  questions: GetAllQuestion;
  //private
}

export interface GetAllAnswerAndQuestion {
  nongsAnswers: UserAndAllQuestionPack[];
  peeAnswers: UserAndAllQuestionPack[];
  mainChoices: InterChoiceQuestion[];
  mainTexts: InterTextQuestion[];
  nongPendingAnswers: UserAndAllQuestionPack[]; /////////////i
  nongPassAnswers: UserAndAllQuestionPack[];
  nongSureAnswers: UserAndAllQuestionPack[];
  nongPaidAnswers: UserAndAllQuestionPack[];
  nongInterviewAnswers: UserAndAllQuestionPack[];
  success: boolean;
  camp: BasicCamp;
  //public
}
export interface ScoreTextQuestion {
  id: Id | null;
  score: number;
  //private
}
export interface ScoreTextQuestions {
  scores: ScoreTextQuestion[][];
  campId: Id;
  //public
}
// ...
export interface BasicUser {
  _id: Id;
  name: string;
  lastname: string;
  nickname: string;
  gender: "Male" | "Female";
  shirtSize: Size;
  haveBottle: boolean;
  mode: Mode;
  group: Group | null;
  role: Role;
  fridayActEn: boolean;
  fridayAuth: boolean;
  likeToSleepAtCamp: boolean;
  //private
}
export interface BasicBaan {
  name: string;
  fullName: string | null;
  campId: Id;
  link: string | null;
  _id: Id;
  mdTime: Date;
  nongSendMessage: boolean;
  songIds: Id[];
  canReadMirror: boolean;
  canWriteMirror: boolean;
  mirrorIds: Id[];
  nongIds: Id[];
  peeIds: Id[];
  //public
}
export interface BasicCamp {
  nameId: Id;
  round: number;
  dateStart: Date;
  dateEnd: Date;
  nongDataLock: boolean;
  open: boolean;
  songIds: Id[];
  link: string | null;
  allDone: boolean;
  lockChangePickup: boolean;
  pictureUrls: string[];
  registerModel: "noPaid" | "noInterview" | "all";
  memberStructure:
    | "nong->highSchool,pee->1year,peto->2upYear"
    | "nong->highSchool,pee->2upYear"
    | "nong->1year,pee->2upYear"
    | "nong->highSchool,pee->allYear"
    | "allYearMix";
  logoUrl: string | null;
  registerSheetLink: string | null;
  peeLock: boolean;
  _id: Id;
  campName: string;
  nongSleepModel:
    | "นอนทุกคน"
    | "เลือกได้ว่าจะค้างคืนหรือไม่"
    | "ไม่มีการค้างคืน";
  peeSleepModel: "นอนทุกคน" | "เลือกได้ว่าจะค้างคืนหรือไม่" | "ไม่มีการค้างคืน";
  groupName: string;
  peeDataLock: boolean;
  petoDataLock: boolean;
  haveCloth: boolean;
  actionPlanOffset: number;
  mdTime: Date;
  showCorrectAnswerAndScore: boolean;
  canAnswerTheQuestion: boolean;
  mealIds: Id[];
  canNongSeeAllAnswer: boolean;
  canNongSeeAllActionPlan: boolean;
  canNongSeeAllTrackingSheet: boolean;
  canNongAccessDataWithRoleNong: boolean;
  lockChangeQuestion: boolean;
  canReadTimeOnMirror: boolean;
  nongCall: string;
  boyZoneLadyZoneState: BoyZoneLadyZoneState;
  canNongSeeBaanOrder: boolean;
  nongIds: Id[];
  peeIds: Id[];
  petoIds: Id[];
  boardIds: Id[];
  baanIds: Id[];
  partIds: Id[];
  //public
}
// ...
export type BoyZoneLadyZoneState = (typeof boyZoneLadyZoneStates)[number];